+++ Assets/WorldGeneration/_Scripts/ScriptableObjects/PlantSettings.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using InGameTime;
 using UnityEngine;
 
@@ -6,6 +7,10 @@ namespace WorldGeneration._Scripts.ScriptableObjects
     [CreateAssetMenu(menuName = "ScriptableObjects/Settings/Plant")]
     public class PlantSettings : ScriptableObject
     {
+        public GameObject plantPrefab;
         public InGameDate lifespan;
+        public int maxNumber;
+
+        [HideInInspector] public List<GameObject> plants;
     }
 }