+++ Assets/WorldGeneration/_Scripts/Spawning/BurrowSpawner/BurrowSpawner.cs
@@ -1,7 +1,0 @@
-namespace WorldGeneration._Scripts.Spawning.BurrowSpawner
-{
-    public abstract class BurrowSpawner : AssetSpawner
-    {
-        public bool IsOccupied;
-    }
-}