+++ Assets/WorldGeneration/_Scripts/WorldManager.cs
@@ -14,25 +14,30 @@ namespace WorldGeneration._Scripts
         private Vector2Int resolution = new(128, 128);
 
         [SerializeField] private float maxTerrainHeight = 180.0f;
-        [Range(0.0f, 1.0f)] [SerializeField] private float waterLevel = 0.3f;
+
+        [Tooltip("The height of the water level as a percentage of the maximum terrain height.")]
+        [Range(0.0f, 1.0f)]
+        [SerializeField] private float waterLevel = 0.3f;
+
         [SerializeField] private GeneralSettings generalSettings;
         [SerializeField] private NoiseSettings noiseSettings;
 
         [Header("Objects")] [SerializeField] private GameObject wallPrefab;
-
-        [SerializeField] private GameObject demoCarrot;
-        [SerializeField] private int maxNumberOfCarrots = 20;
-        private AssetManager _assetManager;
-        private NoiseWithClamp _noiseWithClamp;
-
-        private float[,] _map;
-        private Vector2[] _corners;
-
-        private bool _running;
-
+        [SerializeField] private PlantSettings grassSettings;
+        [SerializeField] private PlantSettings carrotSettings;
+        [SerializeField] private Transform grass;
+        [SerializeField] private Transform carrots;
+        
         // private
         private TerrainManager _terrainManager;
         private TimeManager _timer;
+        private AssetManager _assetManager;
+        private NoiseWithClamp _noiseWithClamp;
+        
+        private Vector2[] _corners;
+        private float[,] _map;
+
+        private bool _running;
 
         // Start is called before the first frame update
         private void Start()
@@ -49,8 +54,9 @@ namespace WorldGeneration._Scripts
 
             // _corners = GeneratorFunctions.GetCornerPoints(generalSettings, noiseSettings, _noiseWithClamp);
 
-            _assetManager.SpawnAssets(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, _map,
-                _noiseWithClamp, demoCarrot, _corners, maxNumberOfCarrots);
+            _assetManager.SpawnAssets(resolution, maxTerrainHeight, generalSettings, _map, carrotSettings,
+                carrots,
+                grassSettings, grass);
 
             _running = true;
         }
@@ -79,13 +85,14 @@ namespace WorldGeneration._Scripts
             // Reload Terrain
             if (Input.GetKey(KeyCode.R) && _running)
             {
-                _terrainManager.GenerateTerrain(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings,
+                _terrainManager.GenerateTerrain(resolution, maxTerrainHeight, waterLevel, generalSettings,
+                    noiseSettings,
                     wallPrefab, _noiseWithClamp, out _map);
 
                 // _corners = GeneratorFunctions.GetCornerPoints(generalSettings, noiseSettings, _noiseWithClamp);
 
-                _assetManager.SpawnAssets(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, _map,
-                    _noiseWithClamp, demoCarrot, _corners, maxNumberOfCarrots);
+                _assetManager.SpawnAssets(resolution, maxTerrainHeight, generalSettings, _map,
+                    carrotSettings, carrots, grassSettings, grass);
             }
         }
     }